generator/requirements.txt
fastapi
uvicorn
ollama
Pillow
playwright